fix(sql): add casts for create table as select

+++ core/src/main/java/io/questdb/griffin/engine/functions/cast/CastCharToShortFunctionFactory.java
@@ -41,7 +41,8 @@ public class CastCharToShortFunctionFactory implements FunctionFactory {
     }
 
     @Override
-    public Function newInstance(int position, ObjList<Function> args, IntList argPositions, CairoConfiguration configuration, SqlExecutionContext sqlExecutionContext) {
+    public Function newInstance(int position, ObjList<Function> args, IntList argPositions,
+            CairoConfiguration configuration, SqlExecutionContext sqlExecutionContext) {
         return new Func(args.getQuick(0));
     }
 
@@ -55,9 +56,10 @@ public class CastCharToShortFunctionFactory implements FunctionFactory {
             char c = arg.getChar(rec);
             final byte v = (byte) (c - '0');
             if (v > -1 && v < 10) {
-                return v;
+                return (short) v;
             }
             throw ImplicitCastException.inconvertibleValue(c, ColumnType.CHAR, ColumnType.SHORT);
+
         }
     }
 }
